Fix jumping back from the bsd4grml loader (ldbsd.com) to GNU GRUB
Formerly, we hard-coded jumping to either grub1 (grub-legacy) or
grub2 in the menu option #6, and offered the macros “grub1” and
“grub2” in boot.cfg (the text mode boot prompt / ldbsd.com command
line), with an alias from “grml” to “grub2” (normally).

Now we use the grml-live framework to patching the ldbsd.com con-
figuration files at image creation time to allow for either grub2,
grub1 with stage2, or grub1 with stage2_eltorito to be found.

Note that grub2 will be taken as default iff it exists, because
it supports manifold-boot™ images, which grub-legacy doesn’t.

Signed-off-by: Thorsten Glaser <tg@mirbsd.org>