Pad official ISOs to multiples of 256 KiB.
Add “-no-pad” to all mkisofs calls (the mkisofs padding for CD-ROM
readahead support is 32K or 48K usually), and apply padding ourselves
for partition table support. We will use cylinders of 256 KiB (split
in 16 tracks à 32 sectors à 512 bytes) for these. This means that the
padding we add is 512 bytes at least, and a whole 256 KiB at most.

CD-ROM readahead padding is done by the “-pad” option of cdrecord,
which is enabled by default and adds another 300 KiB, but only on
the CD itself.

Signed-off-by: Thorsten Glaser <tg@mirbsd.org>