templates/boot/grub/grub.cfg: drop insmod part_{msdos,gpt} as it breaks booting
If the modules part_msdos and part_gpt are insmod-ed then booting
with grub (manifold ISO boot) fails with "unaligned pointer
0x........'. The issue can be reproduced with either real™
hardware or using qemu/kvm via restricting memory to ~64MB (using
'-m 64') but won't be present if using something like 512MB
memory in the virtual machine.

Thanks to Axel Beckert for a bugreport and to Thorsten Glaser for
locating the source of the bug.

Testing: issue940
Thanks: Thorsten Glaser <tg@mirbsd.org>