buildd: support IGNORE class to ignore requests to install software packages from other classes
If packages fail to install in autobuilds it's annoying to have
build errors. Even worse it's not having daily ISOs at all if
unstable is broken.

Therefore support the class "IGNORE" in which you can list
all the packages you don't want to install even though the
package(s) might be requested in other classes. Using the
"keyword-" unmarks the package for installation and therefore
we can safely skip packages on daily builds while they are known
to be broken.

Usage example how to ignore speechd-el and ekeyd:

PACKAGES aptitude

speechd-el-
ekeyd-
EOF