Revert "Pass 1:0x96 instead of 4:0x96 to bootilnx.mksh to use first partition as boot system."
This reverts commit aac340b028d5fa45a5e34d07fbd792858607b791.

The switch broke usb keys on some BIOS versions, causing freezes
while detecting the usb stick (without the option to even enter
the BIOS setup). 2010.12 used to work fine and when reverting it
via:

  (dd if=foo.iso bs=1 count=48 skip=462; dd if=foo.iso bs=1 count=16 skip=446) >x;
  dd of=foo.iso bs=1 conv=notrunc count=64 seek=446 if=x

it's working again. So revert to previous state.

Thanks to ryx for reporting and testing and Thorsten Glaser
for his valuable comments.