GRMLBASE.var: set maximum number of packages installed at a time to 3000.
If we don't install all packages at once the packages are
installed in batches of 99 packages at a time. Then packages
which are requested for installation (via "foo" in a package
config file) but fail to install can't be marked for
uninstallation/ignoring (via "foo-") in classes (like IGNORE for
buildd) because the "foo-" might be in a later batch whereas the
failing installation runs in an earlier batch.

Hopefully this doesn't require to much memory usage or breaks
anything, but all the tests so far looked promising.