Support running scripts in bootoption "services" in foreground if they contain the string DO_NO_RUN_IN_BACKGROUND.
If a script configures the system and has to make sure
it's in a consistent state the (custom) init script
shouldn't run as background job but should be able
to block the boot process until it's in a reliable
state.

Actually we might also want to provide a more generic
solution as well, but until we come up with a decent
way this non-intrusive solution solves the problem
at least for custom init scripts.