This patch fixes the incorrect keyboard layout under Xorg.
If ~/.fluxbox/startup is not present startfluxbox will create that file.
The generated ~/.fluxbox/startup runs per default xmodmap ~/.Xmodmap to
configure the keyboard. This clashes with grmls keyboard setup. To prevent
that we add a simplified ~/.fluxbox/startup file which does not run
xmodmap.

Thank to Malte Simon and Sedat Dilek for reporting the wrong keyboard layout
within X11.