Export environment variables for submenues.
Wit Grml 2011.12 we started to use submenus. Unfortunately enivornment
variables must be exported to be used within submenues. This means the
advanced bootoptions in GRUB do not honor loopback or the theme enviornmet
variable.  This patch explictly exports all set environment variables.

Thanks to Michael Schierl for debugging and reporting the issue.