Package validation: also detect packages that failed to install
So far we detected only packages that have been requested for
installation but failed because the package isn't *available*
(as in not available in the configured Debian repositories).

Without checking for packages that have been requested for
installation but failed because of dependency issues we
might be left in a state with missing packages.

So let's check for both situations and generate according JUnit
reports.