Set root password to specified/generated ssh password.
When using Grml as a rescue distribution for dedicated server it is
typically necessary to change the root password manually as user
expect to be able to login into the rescue system directly as root.

This patch modified the behaviour of the ssh bootoption and will set
the root password to the same value as for the grml user.