No longer execute hostname command if hostname option is not present
There's no need to execute the hostname command if the boot
option hostname is not present as /etc/init.d/hostname.sh is
executed on Grml during bootup. We keep the config_hostname()
only because we support setting it to a random hostname if no
argument was specified to the boot option and to keep using
grml-hostname as main interface for changing hostnames
on-the-fly.