Do not accept any command line arguments we can't handle
When executing something like:

| grml-live [...] -V B -q

this doesn't skip the mksquashfs stage (and neither the chroot
build stage of course) because the "B" can't be processed for
obvious reasons.  The user mistyped the "-B" here, but then we
shouldn't even try to execute the command line, so just exit in
such a case.