dpkg-divert grub-probe binary to avoid problems in OpenVZ environments
It's a hard to debug problem as it doesn't seem to be
reproducible in every OpenVZ environment and on the other hand
makes the system unusable as soon as the bug is triggered. :(

On the command line/in the logs you'll notice something like
that before the system is becoming unuseable:

| Setting up grub-common (1.99-22.1) ...
| Setting up grub2-common (1.99-22.1) ...
| Setting up grub-pc-bin (1.99-22.1) ...
| Setting up grub-pc (1.99-22.1) ...
|
| Creating config file /etc/default/grub with new version
| grub-probe: error: cannot find a device for / (is /dev mounted?).
| grub-probe: error: cannot find a device for /boot (is /dev mounted?).
| grub-probe: error: cannot find a device for /boot/grub (is /dev mounted?).

To avoid grub-probe being executed during software installation
let's dpkg-divert the binary during the FAI dirinstall process
through the existing hook script instsoft.GRMLBASE. Then revert
the dpkg-divert via scripts/GRMLBASE/98-clean-chroot before
actually building the ISO again.

While at it also divert update-grub in the same code path as
well.