SW: drop febootstrap from GRML_FULL
While febootstrap was a "tool for bootstrapping a Fedora system
(like Debian debootstrap)" until and including the squeeze
release, starting with wheezy it's a "tool for building supermin
appliances" (actually being a transitional package, depending on
package supermin). It's actually a rewrite of febootstrap but
with its present feature set it doesn't seem to be of any use for
Grml.