Drop .sp workaround, fix some typos in manpages and remove duplicated text
Manpage was including the text .5v at the end of some paragraphs. This
was caused by an old hack in the Makefile which removed the text .sp,
which was accidently added to the end of some paragraphs. (sed -i
's/\.sp//') Looks like this hack is not needed any longer, when
disabling it the .5v disappears as well (which might have been the text
".sp 5v")