Make the GRUB images modular
Now, both the i386-pc and the amd64 UEFI GRUB images are modular.
This means, that startup is faster and that all modules (even those
that cannot be embedded) can be used.

The embedded configfile has been simplified since conditionals are not
supported by the simple command parser of the rescue shell, without
sacrificing any functionality.

Also, the main build script now copies the *.mod and *.lst files for
UEFI GRUB to the ISO.

[My statement that the .lst files are no longer needed was wrong; at
least the command.lst is used for autoloading modules that have
different names than their commands.]