Work around regression in GRUB 2.00-22
Seems that Grub 2.00-22 cannot properly load a gzipped initrd file
with 'linux16' command (which is used to load the gzipped FreeDOS
memdisk image). As a workaround, mount the filesystem using the
'loopback' command (which transparently decompresses it) and then use
blocklist syntax to load the whole mounted filesystem as an initrd.