Provide GRMLBASE script 49-sshd to adjust sshd configuration [Closes: issue1304]
Providing the openssh-server/permit-root-login debconf option sadly
isn't enough to switch from "PermitRootLogin without-password"
to "PermitRootLogin yes", so we have to adjust the file on-the-fly.
(Let's try to avoid shipping the sshd configuration file for now,
to make sure the file works for all the Debian releases.)

While at it also disable UseDNS which speeds up login when DNS
is unconfigured/unavailable/broken.