Do our best to make sure aptitude is installed in chroot
There seems to be a change in Debian (jessie) causing aptitude to
no longer be present in debootstrapped systems by default any
longer. By setting $FAI_DEBOOTSTRAP_OPTS accordingly we include
it in its default package selection though. If it still shouldn't
be present we explicitely try to install it through the
updatebase.GRMLBASE hook script.