Work around "udev >=208-6 without systemd" bug with loopback device, causing serious boot delay
Without the loopback device properly set up when the udev
init script gets executed we have a boot delay of XX seconds.
This is a show stopper for us and until
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=754987
has been resolved we need a workaround for that.

Thanks: Evgeni Golov, Marco d'Itri, Christian Hofstaedtler and Ulrich Dangel for debugging assistance