Fix creation of images based on Wheezy after the udev fix in grml-live
The grml-live repo currently ships an own udev for a bugfix, which only affects
Jessie and later. This version of udev is not installable in Wheezy and results
in an failed image build. Let's pin udev to the Wheezy version on Wheezy, as
the fix is not needed there anyways.