SW: add syslinux-utils to GRMLBASE
Package syslinux-utils added to package config GRMLBASE to support
booting with memdisk. The package includes the binary memdiskfind which
is needed during the boot process.

NOTE from Mika: syslinux-utils is available only since jessie
though building anything older might not be worth troubles unless
someone is actually interested in this, so accept this change as-is.