Enable and disable services by using systemd.presets
By using systemd.presets we can implement our own policy
which services needs to be enabled or disabled on boot-time.

This issue was initially triggered by a problem with the
wpa_supplicant@*.service which failed on boot with an
"Dependency failed for WPA supplicant daemon
(interface-specific version)" error.

See: http://paste.grml.org/2862/