deborphan: add qemu-kvm to list of packages which are never to be reported
In commit ec12218051e8c1 we added qemu-system-x86 as underlying
dependency for qemu-kvm, now qemu-kvm is considered for removal.
This seems to be the case because qemu-kvm depends on
qemu-system-x86 and and even though we explicitely ask for both
packages to be installed, then qemu-kvm is considered for
removal. Let's avoid this by adding qemu-kvm to the list of
packages which are never to be reported by deborphan, then it's
not automatically removed via DEBORPHAN/98-clean-chroot.

Noticed via "grml-live-missing-packages.test_missing_packages_qemu-kvm"
in Jenkins daily builds.