Drop config_stats
The stats feature was integrated to collect some basic
information about usage of Grml. While we anonymise
the data we also lack the according information *why*
things are used in a way they are used (e.g. since
we cut of any arguments to bootoptions we don't
know the actual settings for e.g. vga=xxx).

What we learned from the stats is that grml32 is still
in usage way more often than expected, though a leading
majority of the systems support 64bit, so it's unclear
why those users still use the 32bit version.

While we thought it was a good idea back then we think
different nowadays and users of Grml don't expect such
a behavior as default. So don't make it even opt-in but
just drop it overall.