TTY6/VT6 should start a shell if bootoption startx is not given
When the bootoption "startx" is given, the helper-script
"/etc/init.d/startx" is created during the bootprocess (via
grml-autoconfig).

That means, that if the booptions "startx" is not given, the
helper-script is not created and tty6/vt6 should not try to start X but
should just start a grml-user-shell.

Related to grml/grml#20