Get rid of ugly startx helper script and start X on tty7/vt7
The startx helper script was used to get around some permission problems
which were caused by grml-runtty. As we switched to agetty, we also can
get rid of this ugly hack.

Furthermore we decided to start X on tty7/vt7 (if the bootoption startx
is given) as tty6 "is especially reserved for auto-spawned gettys":

 This is done in order to ensure that there's always a way to get a text
 login, even if due to fast user switching X took possession of more
 than 5 VTs.

See: http://0pointer.de/blog/projects/serial-console.html

Relates to grml/grml#20