templates: fix isolinux boot option label for grub
A long time ago (rev# 7d0f02a63) grub2 was introduced and the boot option
label grub was changed to grub1 (aside with grub2). Later grub1 support
was removed (rev# bfbf9fb3). So it makes sense to change the grub2 label
to grub as we do not have anything else anymore.