brltty: start brltty service instead of invoking /lib/brltty/brltty.sh
The /lib/brltty/brltty.sh script doesn't seem to work as such any longer
to start the service.  When starting brltty via `systemctl start brltty`
it at least starts (I'm not sure it's actually working for users of
braille devices., but don't know how to verify it on my own).

Closes: https://github.com/grml/grml/issues/58