Set up Unicode/UTF8 via console-setup
This change fixes a couple of problems where strange characters were
shown at the console on input (like when typing german umlauts) and on
output (e.g. `pstree`).

We tried to fix that by changing the order in which we set up the fonts,
then run loadkeys and finally invoked unicode_start via grml-autoconfig
(see: grml/grml-autoconfig@c820a66). But this only changed the behavior
on tty1, the other consoles still had problems when trying to display
unicode characters.

I believe that the "activate unicode console if running within utf8
environment" code in the config_language() function in grml-autoconfig
is not needed after applying this change (i.e. we then do not have to run
unicode_start in grml-autoconfig anymore). But this needs testing.

One more note regarding the "Requires:"-line in `grml-boot.target`: I am
not quite sure why I had to add `console-setup.service` in there,
although the service is enabled via `10-grml.preset`. But without adding
`console-setup.service` to the "Requires:"-line the service is not
started on boot. I suspect this has to do with the fact that the service
has `WantedBy=multi-user.target` in the `[Install]` section, but so has
`rsyslog.service` which starts fine (the other enabled services have
other dependencies). This needs investiagation.

Thanks for feedback: Michael Schierl (@schierlm) and @qlplq
Relates to: grml/grml#101
Closes: grml/grml-autoconfig#9, grml/grml#50