config_language: rely on console-setup for keyboard + font handling
Our old approach with running loadkeys, setting console font and
invoking unicode_start via grml-autoconfig is incomplete for nowadays'
environments.

We tried to fix that by changing the order in which we set up the fonts,
runnning loadkeys and finally invoking unicode_start (see commit
497558a23c). But this changed only the behavior on tty1, the other
consoles still had problems when trying to display unicode characters
(see e.g. `systemctl status` output).

Instead when running under systemd rely on console-setup, by assuming
/etc/default/console-setup is set up as needed (implemented in grml-live
>=0.33.4) and adjusting /etc/default/keyboard according to
keyboard=.../lang=... boot options. Finally we invoke the console-setup
service.  (We might want to handle this outside of grml-autoconfig in
the future, but this approach for now guarantees to execute this in the
right order and not spit on the console at the end of the boot process).

Closes: grml/grml-autoconfig#9, grml/grml#50
Thanks: Darshaka Pathirana for debugging this