/etc/skel/.xinitrc.d/grml_keyboard: support /etc/default/keyboard usage
Starting with Grml 2018.12 we try to rely on console-setup and
keyboard-configuration packages as much as possible.  We no longer want
to generate nor support file /etc/sysconfig/keyboard any longer (while
current grml-lang / grml-setkeyboard / grml-setlang scripts are known to
still use it), but we stay backwards compatible with this change.

Related to grml/grml-autoconfig#9 and grml/grml#50