lvm: start lvm2-lvmetad only if present + support lvm2-lvmpolld
lvm2-lvmetad is gone since 2.03.01-1 (and its init script
was actually dropped later in 2.03.02-2).

Instead lvmpolld is available since 2.02.122-1.

While at it improve formatting, as startup of lvm2-lvmpolld (and
lvm2-lvmetad if present) displays a message on initial startup which
gets in between the "Searching for logical volumes and enabling them:"
and the display of present LVM devices, which isn't nice. So properly
split them.

Closes: grml/grml-autoconfig#10