grml-live: fix falldown from c78f58bfee79af61276e3e91457a1b0faa55bbf5.
While deleting the bind-mount destination of mirror directories is fine
in theory, we ended up deleting empty base directories as well.

This is obviously bad, so let's avoid removing base directories by:
  - keeping a copy of all base directories (this should be updated
    periodically, but updates to base directories happen so rarely that
    it hardly matters)
  - wrapping the functionality of rmdir -p in a new function that makes
    sure to stop on base directories, no matter whether empty or not.

Since we use that functionality for directories in a chroot (from the
outside), additional code to handle chroots is necessary.

With that, base directories should now be correctly protected from
removal.