Revert "grml-live: clean up MIRROR_DIRECTORY if provided."
This reverts commit c78f58bfee79af61276e3e91457a1b0faa55bbf5.

While deleting the bind-mount destination of mirror directories is fine
in theory, we end up deleting empty base directories as well, which
is unexpected and unwanted behavior.

See https://github.com/grml/grml-live/pull/104 for the related
discussion, this issue needs further attention yet before we can
include this in a new release of grml-live.