Update DEBIAN_STABLE's apt sources.list configuration for bullseye
Related fix for commit 07181b4c86f. Now that bullseye is the new Debian
stable release, also adjust the backports entry accordingly.

FTR, when using "stable-backports" instead of "bullseye-backports",
apt warns about:

| W: Conflicting distribution: http://ftp.debian.org/debian stable-backports InRelease (expected stable-backports but got bullseye-backports)

We might not worry about this, but it might be confusing for users,
so until we've a better approach for this, let's keep it as-is.

Thanks: d630 for the bug report
Closes: https://github.com/grml/grml-live/issues/105