SW: drop bacula-* from GRML_FULL
Bacula was removed from Debian/testing as of 2021-11-23 due to #997139.
It's unclear, whether bacula is of any further use on a live system
as-is (also given that the bacula version in Debian is lagging behind
upstream quite a lot; and there also exists bareos nowadays, though not
present in Debian as such). If someone should be missing bacula from
Grml, I'd like to know.

Not only drop the software packages, but also drop the hostname
workaround for /etc/bacula/bacula-fd.conf in our GRMLBASE/05-hostname
script.