SW: switch from bsdmainutils to bsdextrautils + bsdutils in GRML_SMALL + GRML_FULL
The bsdmainutils package is a transitional package (at least) since
Debian/bullseye. It depends on bsdutils, debianutils, bsdextrautils +
ncal. We want bsdutils (e.g. for logger) and bsdextrautils (e.g. for
hexdump), so let's depend on those. The debianutils package is a
required and essential package in Debian anyway, so no need to
explicitly depend on it. We don't necessarily need ncal, so at least
until someone asks for it, let's not depend on it.

FTR, not modifying /etc/grml/fai/config/scripts/DEBORPHAN/10-whitelist
yet, to give people building ISOs with grml-live based on Debian
releases older than bullseye the option to still use it as-is.