SW: drop xfsdump from GRML_FULL
xfsdump fails to compile against Linux kernel versions >=5.17, resulting
in a FTBFS RC bug in Debian (see #1013549).

We have failing *-testing ISO builds since 2022-08-09, as xfsdump is no
longer in Debian testing due to this RC bug.

The package is in a pretty bad shape and I stumbled upon too many issues
when looking into a NMU, so let's drop this for now.