grml_wallpaper: support setting wallpaper background with xwallpaper
The fbsetbg binary is available only via the fluxbox package.
We temporarily switched from fluxbox to openbox (as fluxbox wasn't
available in Debian/testing for quite some time), but then we don't
get our wallpaper set as expected. The smallest working tool we could
identify to set a wallpaper background seems to be xwallpaper. If
someone replaces fluxbox with something else, setting the wallpaper is
supported, as longs as xwallpaper is available.

Relatedto https://github.com/grml/grml-desktop/issues/5