Increase EFI image size when using Secure Boot on amd64
Now as of git commit 721a473 the disk usage of the Secure Boot files
increased and 4MB are no longer enough. We don't support Secure Boot
for i386, so when either building for i386 or amd64 without Secure Boot
then keep the default 4MB size. But if Secure Boot is not disabled
and when not building for i386 (:= amd64) then increase EFI size to 8MB.

Thanks: János Pásztor for bugreport and initial patch