Replace deprecated vmware-detect with virt-what/imvirt
As of git commit 0c1cd5d0cbee in grml-scripts we no longer ship our
vmware-detect binary, so instead rely on virt-what and imvirt.

Output of virt-what executed inside a VM on a VMware cluster is `vmware`
and `VMware ESX server` for imvirt.

Thanks: Christoph Biedl and András Korn for feedback