.Xresources: Add Copy/Paste Key Binding
On systems without a middle mouse button it is not always obvious how to
paste the selection into an xterm window.

For our convenience I added the custom key binding Ctrl-Shift-v +
Ctrl-Shift-c to make our lives a little less annoying.

We also lack documentation on what is currently configured and how to
change it. In an attempt to improve the situation, I have added some
online links and some detailed comments on the lines I have added.

Note, that the last escaped newline (\n\) is not needed, but makes it
easier to add more key bindings.

Closes: grml/grml-desktop#6