GRMLBASE/91-update-pciids: also ignore exit code 1
If the grml_chroot uses /etc/resolv.conf with a symlink pointing to
/run/resolvconf/resolv.conf which might not exist, executing
update-pciids fails.

In the past, wget used to fail with exit code 4 (`Network failure`
according to wget(1)). Nowadays we seem to get exit code 1 (`Generic
error code`) instead. It behaves as such even if curl is present and
gets prefered over wget.

So let's also catch this error condition, to not fail a build because of
that, but instead warn about it.