Add debootstrap to Depends
As of git rev d6d5fee3 we no longer depend on fai-server, though
fai-client doesn't depend on debootstrap, so we need to track this
on our own.

Would be nice to be able to use mmdebstrap instead of debootstrap,
but sadly FAI has this hardcoded, so until we got rid of FAI, we
need to explicitly depend on debootstrap.